requirements.txt
numpy
Pillow
torch 